+++ Assets/StockMarket/Events/MoneyEvent.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "New MoneyEvent", menuName = "Event/MoneyEvent")]
 public class MoneyEvent : Event
 {
-    public int money;
+    public long money;
     public override void Act(Game game){
         game.cash += money;
     }
